requirements.txt
django
psycopg2
pillow
django-cleanup
djangorestframework
djangorestframework-gis